Initial restructure of models file into a module.
